Move PR concepts before file structure
The original order didn't really make sense.
First I've explained the fork and PR model in the context of
'normal' team collaboration.
Then we discuss the lesson files and how to build the site locally.

Finally with all this information, we can make PR.
Exercises first on building with Jekyll, then submitting a PR.